BIG BUG removed.
The problem was that function remove which was supposed to
erase an element from the list inside a loop (thought its iterator)
was not working propery, overstepping some elementes or calling die()
of some elements more than onces. This bug made a big impact on the
dynamics of the tree.